# perf(ext/web): optimize for single pass utf8 decoding