This commit implements a fast path for
decoding common non-streaming utf8
input data.

Benchmark results for decoding a 145MB 
CSV file using 1024 bytes as chunk size:
```
# This patch
{
  opsDispatched: 18635836,
  opsDispatchedSync: 18500048,
  opsDispatchedAsync: 135788,
  opsDispatchedAsyncUnref: 0,
  opsCompleted: 18635836,
  opsCompletedSync: 18500048,
  opsCompletedAsync: 135788,
  opsCompletedAsyncUnref: 0,
  bytesSentControl: 0,
  bytesSentData: 0,
  bytesReceived: 0
}
Read 500001 lines for 8.201 seconds

# Deno 1.23.0
{
  opsDispatched: 18635836,
  opsDispatchedSync: 18500048,
  opsDispatchedAsync: 135788,
  opsDispatchedAsyncUnref: 0,
  opsCompleted: 18635836,
  opsCompletedSync: 18500048,
  opsCompletedAsync: 135788,
  opsCompletedAsyncUnref: 0,
  bytesSentControl: 0,
  bytesSentData: 0,
  bytesReceived: 0
}
Read 500001 lines for 16.362 seconds
```

1024 * 1024 bytes chunk size:
```
# This patch
{
  opsDispatched: 18500183,
  opsDispatchedSync: 18500048,
  opsDispatchedAsync: 135,
  opsDispatchedAsyncUnref: 0,
  opsCompleted: 18500183,
  opsCompletedSync: 18500048,
  opsCompletedAsync: 135,
  opsCompletedAsyncUnref: 0,
  bytesSentControl: 0,
  bytesSentData: 0,
  bytesReceived: 0
}
Read 500001 lines for 11.447 seconds

# Deno 1.23.0
{
  opsDispatched: 18500183,
  opsDispatchedSync: 18500048,
  opsDispatchedAsync: 135,
  opsDispatchedAsyncUnref: 0,
  opsCompleted: 18500183,
  opsCompletedSync: 18500048,
  opsCompletedAsync: 135,
  opsCompletedAsyncUnref: 0,
  bytesSentControl: 0,
  bytesSentData: 0,
  bytesReceived: 0
}
Read 500001 lines for 20.987 seconds
```

Fixes #14893