@vslinko 2022-06-17
@littledivy Could you please run `make test-node-csv-parse` to compare performance with Node's csv-parse on your computer?