@littledivy 2022-06-17
@vslinko 

```
make test-node-csv-parse
test-node-csv-parse
Read 500001 lines for 6.568 seconds
```