@bartlomieju 2022-06-18
> CSV file using 1024 bytes as chunk size:

Have you tried other chunk sizes? Why 1024 bytes?